+++ tests/test_io_rasterio.py
@@ -102,7 +102,6 @@ def test_info_valid():
 
         assert src.colormap
         meta = src.info()
-        assert meta["colormap"]
         assert meta.colormap
 
     with Reader(COG_NODATA, colormap={1: (0, 0, 0, 0)}) as src:
@@ -371,12 +370,6 @@ def test_statistics():
         assert stats["b1"].percentile_2
         assert stats["b1"].percentile_98
 
-        with pytest.warns(DeprecationWarning):
-            assert stats["b1"]["percentile_2"]
-
-        with pytest.warns(DeprecationWarning):
-            assert stats["b1"]["percentile_98"]
-
     with Reader(COGEO) as src:
         stats = src.statistics(percentiles=[3])
         assert stats["b1"].percentile_3
@@ -425,7 +418,6 @@ def test_Reader_Options():
     with Reader(COGEO, options={"nodata": 1}) as src:
         assert src.info().nodata_value == 1
         assert src.info().nodata_type == "Nodata"
-        assert src.info()["nodata_type"] == "Nodata"
 
     with Reader(COGEO) as src:
         assert src.info().nodata_type == "None"