+++ tests/test_models.py
@@ -237,11 +237,6 @@ def test_point_data():
     for p in PointData(numpy.zeros((3), dtype="uint16")):
         assert p == 0
 
-    pt = PointData(numpy.zeros((3), dtype="uint16"))
-    with pytest.warns(DeprecationWarning):
-        arr = pt.as_masked()
-    assert numpy.ma.is_mask(arr.mask)
-
     pt1 = PointData(numpy.array([1, 2]))
     pt2 = pt1.apply_expression("b1+b2")
     assert pt1.count == 2
@@ -352,40 +347,6 @@ def test_image_apply_colormap():
 
     assert im.array.mask[:, 0, 0].tolist() == [True, True, True]
     assert im.mask[0, 0] == 0
-
-
-def test_image_from_array():
-    """Create ImageData from arrays."""
-    with pytest.warns(DeprecationWarning):
-        arr = numpy.zeros((1, 256, 256), dtype="uint8")
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert im.mask.all()
-
-        arr = numpy.ma.MaskedArray(numpy.zeros((1, 256, 256), dtype="uint8"))
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert im.mask.all()
-
-        arr = numpy.ma.MaskedArray(
-            numpy.zeros((1, 256, 256), dtype="uint8"),
-            mask=numpy.zeros((256, 256), dtype="uint8"),
-        )
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert im.mask.all()
-
-        mask = numpy.zeros((256, 256), dtype="uint8")  # 0 no masked
-        mask[0:10, 0:10] = 1  # masked
-        arr = numpy.ma.MaskedArray(
-            numpy.zeros((1, 256, 256), dtype="uint8"),
-            mask,
-        )
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert not im.mask.all()
-        assert im.mask[0, 0] == 0
-        assert im.mask[11, 11] == 255
 
 
 def test_image_from_bytes():